pyproject.toml
[project]
name = "large-lecture-model"
version = "0.1.0"
description = "AI study assistant for college courses"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "playwright>=1.40",
    "pymupdf>=1.24",
    "webvtt-py>=0.4",
    "spacy>=3.7",
    "sentence-transformers>=3.0",
    "chromadb>=0.5",
    "networkx>=3.3",
    "piazza-api>=0.9",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
